Zero password buffers where possible
We cannot zero the password buffer when passed in on the command-
line, since the arg-parsing portion owns that string (but it was
on the command line already, so I'm not super concerned). And on
output, the password goes into the clipboard, which is out of our
control. But on input, when entered into the credential prompt,
it should be zeroed out from memory completely after it has been
encrypted and put into the vault.