Fix my first real bug!
Clippy told me to go through and remove the `&` in all my match
statements, but that made my pointer point to garbage, because I started
moving my vector bytes instead of referring to them. And the borrow
checker didn't catch it because it was in an unsafe block. Also added a
minimal test case that reproduced the issue.