Pass the `AT_KEYEXCHANGE` flag for the smart card KSP
I have no clue why this is needed, and it took a *lot* of trial and
error to figure out that this fixes the issue. If I take my yubikey
to another machine for the first time, it cannot open the vault key.

I can see that the key is still present, it just can't seem to open
keys on machines that didn't create the key (tested with other vaults).

This smells like a cache, where when cached, this hint is not
required, but on machines without that cache entry, something fails
if this hint is not present. I have no idea, and this stuff is *not*
well documented. Feels pretty hacky.